Add workaround for anonymous user
The bug:
I created an User object with id=0, so when saving this to the
database a second time it causes problem. Because there are
two id=0 user instances.

The workaround:
Never save any user with id=0 to database.

The future plan:
This is not ideal, a better solution is to take advantage of the
anonymous scheme comes with flask.